+++ client/src/components/landingPageComponents/NavBar.jsx
@@ -1,10 +1,11 @@
 import React from "react";
+import UserAccount from "./UserAccountContainer";
 
 const NavBar = () => {
     return (
         <div>
             <img src=".../assets/logo.svg"/>
-            <button>This button will be where the login component is.</button>
+            <UserAccount/>
         </div>
     )
 };
